Removing reference to the webbrowser control for the moment
Had a loading error with it and decided it's best to hold off using it
for the moment until the time calls for it. I may consider using the RTF
control for the built-in help instead, but I would lose hyperlinks. Not
sure what the most efficient and reliable method is for embedding help.