bluWindow: Slightly simplify the form dragging
Also, add warning that the form's `Click` event will no longer fire for
the left mouse button. I played around a bit with return values and
calling the original WndProc, but I couldn't seem to get it to drag the
form as well as pass on the click so that VB would get it. May revisit
this in the future if I ever need a Form_Click event