Drag and drop import of ROM
Does not yet copy the ROM to app data, but we now have the first quarter
of a complete workflow. Starting the application looks for a ROM, if not
present it displays a form to provide the ROM and then continues on to
the main application. (We are now no longer dependent on my particular
named ROM being in a particular location)