Added sizer control to bluControlBox
* Now we have a user control for window sizing handle that automatically
negotiates the MDI parent if necessary. This has made it a _lot_ easier
to put a sizer on each child form, as the MDI form does not have any
shared statusbar. We can now resize the window regardless of which child
form is displayed
* Also, we have set up a statusbar in frmLevel, some of the controls
have been moved there, but are not visible yet until implemented