Changes to installation / data location
* Default local install has been changed from Local App Data to Roaming
App Data. This simplifies a lot of things regarding storing application
data and user data -- essentially the app always behaves as if it's
portable, it writes data to the app path and we no longer have to
attempt to detect if we are locally installed and split off data into
Local / Remote App Data folders. It is perhaps not quite 100% "proper",
but it is infinitely easier. The location of everything is predictable,
and not a tangle of conditions
* The application's data (ROM file, update files) will now be in "\App
Data" and the project files in "\User Data" instead of a single "Data"
folder, with user-data going elsewhere
* Removed, therefore, the checks in the app for the %LOCALAPPDATA%
folder, the installer handles the correct placement and we don't need to
test in the app