First work towards automated updating
This is a very messy half-complete commit, but I had to experiment a bit
to get my ideas together. It's been a bit of a slog pulling this
together, ergo the long wait between commits.

* On start up, if "App Data" doesn't exist it gets created, otherwise
the app fails -- it cannot be run from a read-only / admin-only folder.
TODO: Add logging
* Removed, therefore, "frmROM" creating the "App Data" folder
* ROM can only be in one location now, don't require `ROM.Locate`
* Fixed up and tested the bluDownload control
* Added INIFile class to read from a settings file (incomplete)
* `mdiMain` now checks the INI file for last update check and initiates
a check if needed
* Change use of `""` to `vbNullString`, it's faster / less memory
* Created experimental bluWebView control to embed Internet Explorer
without requiring a project reference (which can be unstable when
distributing code) -- this might get removed later