Release test
I'm looking at perhaps automating releases through GitHub with the
releases API. However, I need to know the URL of attached binaries so
that I can get the app to download the new installer. This may be a
chicken and egg problem. This commit is just to act as a test release to
see what GitHub does and where the binary goes.