Got most of the sprite layouts ripped from the ROM!
This adds a lot of the object graphics from the game to the editor in a
reasonably automated fashion. However there are still many exceptions.
Also there's a lot of odd misalignment I hadn't expected and won't get
to the bottom of until I've disassembled the object code in the ROM.